requirements.txt
pdfminer.six
#tkinter
